implement curd in article

+++ controllers/article.js
@@ -3,28 +3,32 @@ const Article = require('../models/article');
 exports.listArticle = async (req, res, next) =>
 {
     try {
-        const article = await Article.find().populate('author');
-        // const articles = {
-        //     slug: article.slug,
-        //     title: article.title,
-        //     description: article.description,
-        //     body: article.body,
-        //     tagList: article.tagList,
-        //     createdAt: article.createdAt,
-        //     updatedAt: article.updatedAt,
-        //     favorited: article.favorited,
-        //     favoritesCount: article.favoritesCount,
-        //     author: {
-        //         username: article.author.username,
-        //         bio: article.author.bio,
-        //         image: article.author.image,
-        //         following: article.author.following
-        //     }
-        // }
-        // console.log(article.author);
+        const articles = await Article.find({}).populate('author');
 
 
-        res.json({ articles: [article] })
+        const listarticles = articles.map((article) =>
+        {
+            const { username, bio, image, following } = article.author;
+            const { slug, title, description, body, tagList, createdAt, updatedAt, favorited, favoritesCount } = article;
+            const author = { username, bio, image, following };
+            newArticle = {
+                slug,
+                title,
+                description,
+                body,
+                tagList,
+                createdAt,
+                updatedAt,
+                favoritesCount,
+                favorited,
+                author
+            }
+            return newArticle;
+        })
+
+
+
+        res.json({ listarticles });
     } catch (error) {
         next(error);
     }
@@ -104,27 +108,36 @@ exports.updateArticle = async (req, res, next) =>
     try {
         const slug = req.params.slug;
         const data = req.body.article;
-        await Article.findOneAndUpdate({ slug }, data);
+        const currentUserId = req.user.userId;
+        // const article = await Article.findOne({ slug }).populate('author');
         const article = await Article.findOne({ slug }).populate('author');
-        const newArticle = {
-            slug: article.slug,
-            title: article.title,
-            description: article.description,
-            body: article.body,
-            tagList: article.tagList,
-            createdAt: article.createdAt,
-            updatedAt: article.updatedAt,
-            favorited: article.favorited,
-            favoritesCount: article.favoritesCount,
-            author: {
-                username: article.author.username,
-                bio: article.author.bio,
-                image: article.author.image,
-                following: article.author.following
-            }
-        }
+        const authorId = article.author.id;
+        if (currentUserId === authorId) {
 
-        res.json({ newArticle })
+            await Article.findOneAndUpdate({ slug }, data);
+            const article = await Article.findOne({ slug }).populate('author');
+            const newArticle = {
+                slug: article.slug,
+                title: article.title,
+                description: article.description,
+                body: article.body,
+                tagList: article.tagList,
+                createdAt: article.createdAt,
+                updatedAt: article.updatedAt,
+                favorited: article.favorited,
+                favoritesCount: article.favoritesCount,
+                author: {
+                    username: article.author.username,
+                    bio: article.author.bio,
+                    image: article.author.image,
+                    following: article.author.following
+                }
+            }
+
+            res.json({ newArticle })
+        } else {
+            res.json("you can't update this article");
+        }
 
     } catch (error) {
         next(error);
@@ -135,27 +148,19 @@ exports.deleteArticle = async (req, res, next) =>
 {
     try {
         const slug = req.params.slug;
-        await Article.findOneAndDelete({ slug });
-        // const article = await Article.findOne({ slug }).populate('author');
-        // const newArticle = {
-        //     slug: article.slug,
-        //     title: article.title,
-        //     description: article.description,
-        //     body: article.body,
-        //     tagList: article.tagList,
-        //     createdAt: article.createdAt,
-        //     updatedAt: article.updatedAt,
-        //     favorited: article.favorited,
-        //     favoritesCount: article.favoritesCount,
-        //     author: {
-        //         username: article.author.username,
-        //         bio: article.author.bio,
-        //         image: article.author.image,
-        //         following: article.author.following
-        //     }
-        // }
-
-        res.json("article deleted")
+        const currentUserId = req.user.userId;
+        // console.log("from auth userId", userId);
+        // console.log(slug);
+        const article = await Article.findOne({ slug }).populate('author');
+        const author = article.author;
+        const authorId = author.id;
+        // console.log("article userId", author.id);
+        if (currentUserId === authorId) {
+            await Article.findOneAndDelete({ slug });
+            res.json("article deleted");
+        } else {
+            res.json("you can't delete this article");
+        }
 
     } catch (error) {
         next(error);
